refactor: rename enums.dart to lean_types.dart
