+++ lib/lean_web_client.dart
@@ -3,7 +3,7 @@ import 'dart:developer';
 import 'package:url_launcher/url_launcher.dart';
 import 'package:webview_flutter/webview_flutter.dart';
 
-import 'enums.dart';
+import 'lean_types.dart';
 
 class LeanWebClient {
   static LeanResponse _getResponseFromParams(Uri uri) {
